mastitis_detection_server/requirements.txt
Django==4.0
django-cors-headers==3.11.0
djangorestframework==3.12.4